+++ src/commaLists.js
@@ -1,0 +1,7 @@
+import tags from './tags'
+
+export default tags({
+  includeArrays: {
+    separator: ','
+  }
+})
