+++ src/commaListsAnd.js
@@ -1,0 +1,8 @@
+import tags from './tags'
+
+export default tags({
+  includeArrays: {
+    separator: ',',
+    conjunction: 'and'
+  }
+})
