+++ src/oneLineCommaListsOr.js
@@ -1,0 +1,9 @@
+import tags from './tags'
+
+export default tags({
+  includeArrays: {
+    separator: ',',
+    conjunction: 'or'
+  },
+  oneLine: true
+})
