+++ src/stripIndent.js
@@ -1,0 +1,3 @@
+import tags from './tags'
+
+export default tags({ stripIndent: true })
